fixed function prototypes

+++ usr/lib/zero/randmt32.h
@@ -4,8 +4,9 @@
 #include <stdint.h>
 
 #define RANDMT32_MAX 0xffffffff
-void srandmt32(int32_t seed);
-int  randmt32(void);
+
+void          srandmt32(int32_t seed);
+unsigned long randmt32(void);
 
 #endif /* __ZERO_RANDMT32_H__ */
